refactor(md5): user hasher, move cmd.
Close #21
